Extract the neutral task shell into Hardened.Idl.BuildTask
Slicing, diagnostics, emitting and writing are identical whatever language the
description was written in - all of them operate on ServiceSpecModel and none
opens a document. ExtractOpenApiSpec drops from 409 lines to 85: a Parse
override, the three properties only an OpenAPI document has, and the strings
that name this front end in diagnostics.

The extension point costs nothing it was thought to. ExtractOpenApiSpec
documented the seam as needing "an unseal plus a model promoted to public - real
surface, on a shipped MSBuild task", because a protected member returning an
internal type is CS0050. An internal abstract member on a public abstract class
is legal and says the right thing instead: the class can be derived inside this
assembly and nowhere else, which is the only place a front end can live given
the IR is compiled in beside it. No consumer surface widens.

Linked as source rather than referenced, because ServiceSpecModel is internal
and each front end compiles its own copy of the IR. A base class in a third
assembly would name a different type and would not bind.

NameAllocator.Apply deliberately stays in the front end. It has to run after
front-end-specific reshaping - the ordering comment in OpenApiSpecParser says as
much - and roughly forty tests call Parse directly so they exercise what the
task exercises. Moving it would have made all of them assert against a model the
build never produces. It is documented as a contract on Parse instead.

Hardened.Idl.BuildTask is in the solution so it compiles standalone, which is
the guard: it caught CSharpAuthor needing System.Net.Http on net472 immediately,
where a front end would have hit it later and noisier.

2891 tests, unchanged from baseline.
