+++ src/Hardened.Framework.sln
@@ -131,6 +131,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Hardened.IntegrationTests.A
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Hardened.IntegrationTests.Authorization.SUT.Tests", "IntegrationTests\Authorization\Hardened.IntegrationTests.Authorization.SUT.Tests\Hardened.IntegrationTests.Authorization.SUT.Tests.csproj", "{6F8E0271-1E25-4130-9F51-34AF8C538D26}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Hardened.Idl.BuildTask", "SourceGenerators\Hardened.Idl.BuildTask\Hardened.Idl.BuildTask.csproj", "{40A181A6-6BA3-422B-858F-DA9CE5F2E3BC}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -765,6 +767,18 @@ Global
 		{6F8E0271-1E25-4130-9F51-34AF8C538D26}.Release|x64.Build.0 = Release|Any CPU
 		{6F8E0271-1E25-4130-9F51-34AF8C538D26}.Release|x86.ActiveCfg = Release|Any CPU
 		{6F8E0271-1E25-4130-9F51-34AF8C538D26}.Release|x86.Build.0 = Release|Any CPU
+		{40A181A6-6BA3-422B-858F-DA9CE5F2E3BC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{40A181A6-6BA3-422B-858F-DA9CE5F2E3BC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{40A181A6-6BA3-422B-858F-DA9CE5F2E3BC}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{40A181A6-6BA3-422B-858F-DA9CE5F2E3BC}.Debug|x64.Build.0 = Debug|Any CPU
+		{40A181A6-6BA3-422B-858F-DA9CE5F2E3BC}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{40A181A6-6BA3-422B-858F-DA9CE5F2E3BC}.Debug|x86.Build.0 = Debug|Any CPU
+		{40A181A6-6BA3-422B-858F-DA9CE5F2E3BC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{40A181A6-6BA3-422B-858F-DA9CE5F2E3BC}.Release|Any CPU.Build.0 = Release|Any CPU
+		{40A181A6-6BA3-422B-858F-DA9CE5F2E3BC}.Release|x64.ActiveCfg = Release|Any CPU
+		{40A181A6-6BA3-422B-858F-DA9CE5F2E3BC}.Release|x64.Build.0 = Release|Any CPU
+		{40A181A6-6BA3-422B-858F-DA9CE5F2E3BC}.Release|x86.ActiveCfg = Release|Any CPU
+		{40A181A6-6BA3-422B-858F-DA9CE5F2E3BC}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -826,6 +840,7 @@ Global
 		{09C4FE7A-57F1-4CAC-B3A1-2B548C9D1034} = {EABCDDC0-6E72-43FB-9AF7-04E6681B83FD}
 		{397E3503-9C80-4026-9922-E0B11B7ABF4C} = {592F787C-E5FC-4C0C-8F2F-2DC8559C326B}
 		{6F8E0271-1E25-4130-9F51-34AF8C538D26} = {592F787C-E5FC-4C0C-8F2F-2DC8559C326B}
+		{40A181A6-6BA3-422B-858F-DA9CE5F2E3BC} = {EABCDDC0-6E72-43FB-9AF7-04E6681B83FD}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {BF6CD1AB-6B41-427D-BA1B-C4803CF67E97}
